ui: separate json schema-related functions into JsonSchemaHelpers.js
